pyproject.toml
[project]
name        = "storage-proofs"
version     = "0.1.0"
description = "L2 Oracles powered by storage proofs"
authors     = [{ name = "Curve.fi", email = "info@curve.fi" }]

readme          = "README.md"
requires-python = ">=3.10"

dependencies = [
    "vyper>=0.4.0",
    "snekmate==0.1.0",
    "web3>=7.0.0",
    "ruff>=0.9.6",
    "mamushi==0.0.4a3",
    "pytest-xdist>=3.6.1",
    "py-evm==0.10.1b2",  # Update after `pytest .` check
    # placeholders for git dependencies
    "boa-solidity",
    "titanoboa",
]

[tool.uv.sources]
titanoboa    = { git = "https://github.com/vyperlang/titanoboa.git", rev = "19d2a5d7745ac8ed29639257e8d49bfb1bdb366a" }  # Update after `pytest .` check
boa-solidity = { git = "https://github.com/z80dev/boa-solidity", rev = "3e00649f643ce69108f266942c90bffc717b18b7" }  # Update after `pytest .` check

[tool.ruff]
line-length = 100
